Refactor loggers into a logging folder within the handlers directory
