+++ src/commands/moderation/ban.ts
@@ -9,7 +9,7 @@ import {
     userMention
 } from "discord.js";
 
-import ModLoggingHandler from "../../handlers/modLoggingHandler";
+import ModLoggingHandler from "../../handlers/logging/modLoggingHandler";
 
 const modLoggingHandler = new ModLoggingHandler();
