+++ src/commands/moderation/unmute.ts
@@ -10,7 +10,7 @@ import {
     InteractionContextType
 } from "discord.js";
 
-import ModLoggingHandler from "../../handlers/modLoggingHandler";
+import ModLoggingHandler from "../../handlers/logging/modLoggingHandler";
 const modLoggingHandler = new ModLoggingHandler();
 
 export default class UnmuteCommand implements Command {
